feat: 댓글 좋아요 구현 (#135)
* fix: 테스트 오류 수정

* refactor: 불필요한 로직(recentDate) 제거

* feat: 댓글 좋아요 구현

+++ src/test/java/techcourse/fakebook/web/controller/friendship/FriendshipApiControllerTest.java
@@ -68,7 +68,7 @@ class FriendshipApiControllerTest extends ControllerTestHelper {
         then().
                 statusCode(HttpStatus.OK.value()).
                 body("size", greaterThanOrEqualTo(1)).
-                body("id", hasItem(friendId));
+                body("id", hasItem((int) (long) friendId));
     }
 
     @Test
